Deploying grafana and prometheus using helm chart (#1)
* Deploying grafana and prometheus using helm chart

* Adding ServiceMonitor and Service

We are changing the approach used to provide a
label and scraping target for prometheus to discover
the pods in the k8s cluster. The prometheus values.yaml
parameter 'additionalServiceMonitors' was defined
as the label trakwell-jmeter-status which will allow
the service monitor to publish the service to k8s to
discover pods that are labeled with that service name.

---------

Co-authored-by: Rusty Weise <rusty.weise@jamfast.com>